requirements.txt
aiogram~=2.18
aioredis~=2.0
environs~=9.0
psycopg2-binary==2.9.1
pytz
requests
